review: better scoped tests.
