--- tools/serverpod_cli/test/integration/analyzer/dart/endpoint_validation/endpoint_inheritance_test.dart
@@ -1,66 +1,0 @@
-import 'dart:io';
-
-import 'package:serverpod_cli/src/analyzer/dart/definitions.dart';
-import 'package:serverpod_cli/src/analyzer/dart/endpoints_analyzer.dart';
-import 'package:serverpod_cli/src/generator/code_generation_collector.dart';
-import 'package:serverpod_cli/src/test_util/endpoint_validation_helpers.dart';
-import 'package:serverpod_serialization/serverpod_serialization.dart';
-import 'package:test/test.dart';
-import 'package:path/path.dart' as path;
-
-const pathToServerpod = '../../../../../../../../packages/serverpod';
-var testProjectDirectory = Directory(path.joinAll([
-  'test',
-  'integration',
-  'analyzer',
-  'dart',
-  'endpoint_validation',
-  const Uuid().v4(),
-]));
-
-void main() {
-  setUpAll(() async {
-    await createTestEnvironment(testProjectDirectory, pathToServerpod);
-  });
-
-  tearDownAll(() {
-    testProjectDirectory.deleteSync(recursive: true);
-  });
-
-  group('Given a dart class that does not inherit from Endpoint when analyzed',
-      () {
-    var collector = CodeGenerationCollector();
-    var testDirectory =
-        Directory(path.join(testProjectDirectory.path, const Uuid().v4()));
-
-    late List<EndpointDefinition> endpointDefinitions;
-    late EndpointsAnalyzer analyzer;
-    setUpAll(() async {
-      var endpointFile = File(path.join(testDirectory.path, 'endpoint.dart'));
-      endpointFile.createSync(recursive: true);
-      endpointFile.writeAsStringSync('''
-import 'package:serverpod/serverpod.dart';
-
-class ExampleEndpoint {
-  Future<String> hello(Session session, String name) async {
-    return 'Hello, \$name!';
-  }
-}
-''');
-      analyzer = EndpointsAnalyzer(testDirectory);
-      endpointDefinitions = await analyzer.analyze(collector: collector);
-    });
-
-    test('then no parsing errors are reported.', () {
-      expect(analyzer.getErrors(), completion(isEmpty));
-    });
-
-    test('then no validation errors are reported.', () {
-      expect(collector.errors, isEmpty);
-    });
-
-    test('then no endpoint definition are created.', () {
-      expect(endpointDefinitions, isEmpty);
-    });
-  });
-}
